refactor: legacy 코드를 모두 삭제한다.
