+++ android/app/src/main/java/com/foss/foss/data/FakeMatchRepository.kt
@@ -4,7 +4,7 @@ import com.foss.foss.model.Match
 import com.foss.foss.model.MatchType
 import com.foss.foss.model.Nickname
 import com.foss.foss.model.WinDrawLose
-import com.foss.foss.model.legacy.Score
+import com.foss.foss.model.Score
 import com.foss.foss.repository.MatchRepository
 import java.time.LocalDate
