+++ android/app/src/main/java/com/foss/foss/data/FakeRelativeStatRepository.kt
@@ -4,7 +4,7 @@ import com.foss.foss.model.Nickname
 import com.foss.foss.model.RelativeStat
 import com.foss.foss.model.WinDrawLose
 import com.foss.foss.model.WinDrawLoses
-import com.foss.foss.model.legacy.Score
+import com.foss.foss.model.Score
 import com.foss.foss.repository.RelativeStatsRepository
 import java.time.LocalDate
