--- android/app/src/main/java/com/foss/foss/data/legacy/HttpError.kt
@@ -1,21 +1,0 @@
-package com.foss.foss.data.legacy
-
-import retrofit2.Response
-
-sealed class HttpError(
-    val code: Int,
-    override val message: String
-) : Throwable(message) {
-
-    class BadRequest(
-        response: Response<*>
-    ) : HttpError(response.code(), response.message())
-
-    class EmptyBody(
-        response: Response<*>
-    ) : HttpError(response.code(), "")
-
-    class Unknown(
-        response: Response<*>
-    ) : HttpError(response.code(), "${response.code()} : ${response.body()}")
-}
