--- android/app/src/main/java/com/foss/foss/data/legacy/entity/request/MatchRequest.kt
@@ -1,10 +1,0 @@
-package com.foss.foss.data.legacy.entity.request
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MatchRequest(
-    @SerialName("matchId")
-    val matchId: String
-)
