--- android/app/src/main/java/com/foss/foss/data/legacy/entity/request/MaxRankRequest.kt
@@ -1,5 +1,0 @@
-package com.foss.foss.data.legacy.entity.request
-
-data class MaxRankRequest(
-    val userAccessId: String
-)
