--- android/app/src/main/java/com/foss/foss/data/legacy/entity/request/UserRequest.kt
@@ -1,5 +1,0 @@
-package com.foss.foss.data.legacy.entity.request
-
-data class UserRequest(
-    val nickName: String
-)
