--- android/app/src/main/java/com/foss/foss/data/legacy/entity/response/DefenceResponse.kt
@@ -1,16 +1,0 @@
-package com.foss.foss.data.legacy.entity.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class DefenceResponse(
-    @SerialName("blockTry")
-    val blockTry: Int,
-    @SerialName("blockSuccess")
-    val blockSuccess: Int,
-    @SerialName("tackleTry")
-    val tackleTry: Int,
-    @SerialName("tackleSuccess")
-    val tackleSuccess: Int
-)
