--- android/app/src/main/java/com/foss/foss/data/legacy/entity/response/MatchResultResponse.kt
@@ -1,16 +1,0 @@
-package com.foss.foss.data.legacy.entity.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MatchResultResponse(
-    @SerialName("matchId")
-    val matchId: String,
-    @SerialName("matchDate")
-    val matchDate: String,
-    @SerialName("matchType")
-    val matchType: Int,
-    @SerialName("matchInfo")
-    val matchesInfoResponse: List<MatchInfoResponse>
-)
