--- android/app/src/main/java/com/foss/foss/data/legacy/entity/response/MatchesIdResponse.kt
@@ -1,5 +1,0 @@
-package com.foss.foss.data.legacy.entity.response
-
-data class MatchesIdResponse(
-    val matchesId: List<String>
-)
