--- android/app/src/main/java/com/foss/foss/data/legacy/entity/response/MaxRankResponse.kt
@@ -1,5 +1,0 @@
-package com.foss.foss.data.legacy.entity.response
-
-data class MaxRankResponse(
-    val ranks: List<RankResponse>
-)
