--- android/app/src/main/java/com/foss/foss/data/legacy/entity/response/RankResponse.kt
@@ -1,12 +1,0 @@
-package com.foss.foss.data.legacy.entity.response
-
-import kotlinx.serialization.SerialName
-
-data class RankResponse(
-    @SerialName("matchType")
-    val matchType: Int,
-    @SerialName("division")
-    val division: Int,
-    @SerialName("achievementDate")
-    val achievementDate: String
-)
