+++ android/app/src/test/java/MatchFixture.kt
@@ -2,7 +2,7 @@ import com.foss.foss.model.Match
 import com.foss.foss.model.MatchType
 import com.foss.foss.model.Nickname
 import com.foss.foss.model.WinDrawLose
-import com.foss.foss.model.legacy.Score
+import com.foss.foss.model.Score
 import java.time.LocalDate
 
 object MatchFixture {
