+++ android/domain/src/main/java/com/foss/foss/model/RelativeStat.kt
@@ -1,6 +1,5 @@
 package com.foss.foss.model
 
-import com.foss.foss.model.legacy.Score
 import java.time.LocalDate
 
 /**
