--- android/domain/src/main/java/com/foss/foss/model/legacy/Rank.kt
@@ -1,9 +1,0 @@
-package com.foss.foss.model.legacy
-
-import com.foss.foss.model.MatchType
-
-data class Rank(
-    val matchType: MatchType,
-    val division: Int,
-    val achievementDate: String
-)
