--- android/domain/src/main/java/com/foss/foss/model/legacy/User.kt
@@ -1,7 +1,0 @@
-package com.foss.foss.model.legacy
-
-data class User(
-    val accessId: String,
-    val nickname: String,
-    val level: Int
-)
